worker_node/requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-jose[cryptography] # jwt 
passlib[bcrtpt] # password hashing
kombu
redis==5.0.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-dotenv
pytest
jinja2
celery==5.3.6
celery[redis]
oss2 

# reconsturct， 后期要调整，将调度服务器与渲染节点 环境分离
plyfile==0.8.1
torch
numpy